+++ packages/scripts/scripts/utils/configureTypescript.js
@@ -150,19 +150,32 @@ module.exports = {
     }
     if (fs.existsSync(paths.cypress)) {
       eslintRoots.push('cypress');
+      packageConfig.exclude.push('src/**/*.cypress.tsx', 'src/**/*.cypress.ts');
+      const compilerPaths = {
+        // this let's us import cypress files from an absolute path in our component tests
+        '#cypress/*': ['../cypress/*']
+      };
+      packageConfig.compilerOptions.paths = compilerPaths;
       writeTsConfig(
         path.join(paths.cypress, 'tsconfig.json'),
         {
           ...packageConfig,
           extends: '@tablecheck/scripts/tsconfig/base.json',
-          include: ['**/*.ts'],
+          exclude: ['node_modules'],
+          include: [
+            '**/*.ts',
+            '../src/**/*.cypress.tsx',
+            '../src/**/*.cypress.ts',
+            '../src/definitions/**/*.ts'
+          ],
           compilerOptions: {
             baseUrl: path.relative(paths.cypress, path.join(paths.cwd, 'src')),
             lib: ['dom', 'dom.iterable', 'esnext'],
             module: 'esnext',
             target: 'es5',
             noEmit: true,
-            isolatedModules: false
+            isolatedModules: false,
+            paths: compilerPaths
           }
         },
         true
@@ -253,9 +266,14 @@ module.exports = {
       }
     };
     const runnerConfigPath = path.join(paths.cwd, 'tsconfig.json');
-    writeTsConfig(runnerConfigPath, config, true);
 
     if (fs.existsSync(paths.cypress)) {
+      config.exclude.push('src/**/*.cypress.tsx', 'src/**/*.cypress.ts');
+      const compilerPaths = {
+        // this let's us import cypress files from an absolute path in our component tests
+        '#cypress/*': ['../cypress/*']
+      };
+      config.compilerOptions.paths = compilerPaths;
       writeTsConfig(
         path.join(paths.cwd, 'tsconfig.eslint.json'),
         {
@@ -273,7 +291,13 @@ module.exports = {
         path.join(paths.cypress, 'tsconfig.json'),
         {
           ...config,
-          include: ['**/*.ts'],
+          exclude: ['node_modules'],
+          include: [
+            '**/*.ts',
+            '../src/**/*.cypress.tsx',
+            '../src/**/*.cypress.ts',
+            '../src/definitions/**/*.ts'
+          ],
           compilerOptions: {
             ...config.compilerOptions,
             baseUrl: path.relative(paths.cypress, path.join(paths.cwd, 'src')),
@@ -285,6 +309,7 @@ module.exports = {
         true
       );
     }
+    writeTsConfig(runnerConfigPath, config, true);
 
     return runnerConfigPath;
   },